Upgrade dependencies and kubebuilder tooling
This is a complete overhaul of this project as we migrated from
kubebuilder plugin go/v3 to go/v4. Following the migration guide
at https://book.kubebuilder.io/migration/v3vsv4.html. I whish
there was a cleaner upgrade path, with less changes, but the
way all this kubebuilder scafolding works and the lack of time
on my side lead to this one big muddy commit.